Major refactoring of store
