+++ src/components/atoms/Predicate.vue
@@ -3,7 +3,7 @@ import {
   defineComponent,
 } from '@vue/composition-api';
 
-import { searchParameters } from '@/store';
+import { searchParameters } from '@/store/search';
 
 export default defineComponent({
   name: 'Predicate',
